+++ cmd/components-list.go
@@ -2,9 +2,10 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/epiphany-platform/cli/internal/logger"
 
+	"github.com/epiphany-platform/cli/internal/logger"
 	"github.com/epiphany-platform/cli/pkg/repository"
+
 	"github.com/spf13/cobra"
 )
