+++ cmd/environments-import.go
@@ -6,6 +6,7 @@ import (
 	"github.com/epiphany-platform/cli/internal/logger"
 	"github.com/epiphany-platform/cli/pkg/environment"
 	"github.com/epiphany-platform/cli/pkg/promptui"
+
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
