+++ cmd/environments-info.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/epiphany-platform/cli/internal/logger"
 	"github.com/epiphany-platform/cli/pkg/environment"
+
 	"github.com/google/uuid"
 	"github.com/spf13/cobra"
 )
