+++ cmd/environments-new.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"github.com/epiphany-platform/cli/internal/logger"
 	"github.com/epiphany-platform/cli/pkg/promptui"
+
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
