+++ cmd/environments-run.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"errors"
+
 	"github.com/epiphany-platform/cli/internal/logger"
 	"github.com/epiphany-platform/cli/pkg/environment"
 	"github.com/epiphany-platform/cli/pkg/processor"
