+++ cmd/repos-install.go
@@ -2,13 +2,13 @@ package cmd
 
 import (
 	"errors"
-	"github.com/spf13/viper"
+	"regexp"
 
 	"github.com/epiphany-platform/cli/internal/logger"
 	"github.com/epiphany-platform/cli/internal/repository"
-	"regexp"
 
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
 var (
