+++ cmd/repos-list.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/epiphany-platform/cli/internal/logger"
 	"github.com/epiphany-platform/cli/internal/repository"
+
 	"github.com/spf13/cobra"
 )
